+++ Timer.h
@@ -5,8 +5,8 @@ class Timer
 {
 public:
 	Timer();
-	double Mark();
-	double Peek();
+	float Mark();
+	float Peek();
 private:
 	std::chrono::steady_clock::time_point last;
 };
